daemon: subclassing of the connectionpool class to avoid code redundancy between implementations
